#72 - refactor: prepare for total weight
